+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  *   Allow animate() to recieve functions
  *   Fix: Image and Button fixes
  *   Fix: Icons now update when chaning params in IE
+ *   Fix: All icons can now be resized
  *   Fix: When setting a value in blocks views changin from advanced to basic setup won't remove quotes in strings
  *   Fix: Check that = has a value
  *   Fix: Default degrees value in arc() is now undefined